LPS-68197 Rename
