+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-form-evaluator-impl/src/main/java/com/liferay/dynamic/data/mapping/form/evaluator/impl/internal/functions/CallFunction.java
@@ -47,7 +47,7 @@ import java.util.Map;
 /**
  * @author Leonardo Barros
  */
-public class CallFunction extends BasePropertyFunction {
+public class CallFunction extends BaseDDMFormRuleFunction {
 
 	public CallFunction(
 		DDMDataProviderConsumerTracker ddmDataProviderConsumerTracker,
