+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-form-evaluator-impl/src/main/java/com/liferay/dynamic/data/mapping/form/evaluator/impl/internal/functions/SetEnabledFunction.java
@@ -22,7 +22,7 @@ import java.util.Map;
 /**
  * @author Leonardo Barros
  */
-public class SetEnabledFunction extends PropertySetFunction {
+public class SetEnabledFunction extends SetPropertyFunction {
 
 	public SetEnabledFunction(
 		Map<String, List<DDMFormFieldEvaluationResult>>
