+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-form-evaluator-impl/src/main/java/com/liferay/dynamic/data/mapping/form/evaluator/impl/internal/functions/SetInvalidFunction.java
@@ -22,7 +22,7 @@ import java.util.Map;
 /**
  * @author Leonardo Barros
  */
-public class SetInvalidFunction extends BasePropertyFunction {
+public class SetInvalidFunction extends BaseDDMFormRuleFunction {
 
 	public SetInvalidFunction(
 		Map<String, List<DDMFormFieldEvaluationResult>>
