+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-form-evaluator-impl/src/test/java/com/liferay/dynamic/data/mapping/form/evaluator/impl/internal/functions/BaseDDMFormRuleFunctionTest.java
@@ -25,7 +25,7 @@ import java.util.Map;
 /**
  * @author Marcellus Tavares
  */
-public abstract class BasePropertyFunctionTest {
+public abstract class BaseDDMFormRuleFunctionTest {
 
 	protected DDMFormFieldEvaluationResult createDDMFormFieldEvaluationResult(
 		String fieldName, String propertyName, Object propertyValue) {
