+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-form-evaluator-impl/src/test/java/com/liferay/dynamic/data/mapping/form/evaluator/impl/internal/functions/SetInvalidFunctionTest.java
@@ -26,7 +26,7 @@ import org.junit.Test;
 /**
  * @author Leonardo Barros
  */
-public class SetInvalidFunctionTest extends BasePropertyFunctionTest {
+public class SetInvalidFunctionTest extends BaseDDMFormRuleFunctionTest {
 
 	@Test
 	public void testEvaluate() {
